Validations done!

+++ Publicaciones/Publicaciones.Application/Services/EmployeeService.cs
@@ -1,7 +1,10 @@
-﻿using Microsoft.Extensions.Logging;
+﻿using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Logging;
 using Publicaciones.Application.Contract;
 using Publicaciones.Application.Core;
 using Publicaciones.Application.DTO.Employee;
+using Publicaciones.Application.Exeptions;
+using Publicaciones.Application.Validations.ContractValidations;
 using Publicaciones.Domain.Entities;
 using Publicaciones.Infrastructure.Interface;
 using System;
@@ -12,12 +15,18 @@ namespace Publicaciones.Application.Service
     public class EmployeeService : IEmployeeService
     {
         private readonly IEmployeeRepository _employeeRepository;
-        private readonly ILogger<EmployeeService> logger;               
+        private readonly ILogger<EmployeeService> logger;
+        private readonly IConfiguration _configuration;
+        private readonly IEmployeeValidations _employeevalidations;
         public EmployeeService(IEmployeeRepository employeeRepository, 
-                                        ILogger<EmployeeService> logger)
+                                        ILogger<EmployeeService> logger,
+                                        IConfiguration _configuration,
+                                        IEmployeeValidations _employeevalidations)
         {
             _employeeRepository = employeeRepository;
             this.logger = logger;
+            this._configuration = _configuration;
+            this._employeevalidations = _employeevalidations;
                 
         }
 
@@ -27,31 +36,24 @@ namespace Publicaciones.Application.Service
             ServiceResult result = new ServiceResult();
             try
             {
+                result.Data = _employeeRepository.GetEmployeeJob();
+                result.Message = _configuration["EmployeeSuccessMessages:getAllEmployeesSuccessMessage"];
 
-                var employees = _employeeRepository.GetEntities()
-                        .Select(EmpGet => new EmployeeDtoGetAll()
-                        {
-                            EmpID = EmpGet.EmpID,
-                            FirstName = EmpGet.FirstName,
-                            LastName = EmpGet.LastName,
-                            Minit = EmpGet.Minit,
-                            Joblvl = EmpGet.Joblvl,
-                            HireDate = EmpGet.HireDate,
-                            JobID = EmpGet.JobID,
-                            PubID = EmpGet.PubID,
-                            ChangeDate = EmpGet.CreationDate,
-                            ChangeUser = EmpGet.IDCreationUser
-                        });
-                        result.Data = employees;
+            }
+            catch (EmployeeServiceExeption exx)
+            {
+                result.Success = false;
+                result.Message = exx.Message;
+                logger.LogError(result.Message, exx.Message);
             }
             catch (Exception ex)
             {
                 result.Success = false;
-                result.Message = $"An error has ocurred while obtaining the employees";
+                result.Message = _configuration["EmployeeErrorMessage:getAllEmployeessErrorMessage"];
                 logger.LogError(result.Message, ex.ToString());
             }
             return result;
-        }
+        } 
 
         //GETBY-ID
         public ServiceResult GetByID(int ID)
@@ -76,46 +78,62 @@ namespace Publicaciones.Application.Service
                         ChangeUser = EmpGet.IDCreationUser
                     };
                     result.Data = employee;
+                    result.Message = _configuration["EmployeeSuccessMessages:getEmployeeSuccessMessage"];
+
+                }
+                catch (EmployeeServiceExeption exx)
+                {
+                    result.Success = false;
+                    result.Message = exx.Message;
+                    logger.LogError(result.Message, exx.Message);
                 }
                 catch (Exception ex)
                 {
                     result.Success = false;
-                    result.Message = $"An error has ocurred while obtaining the employee";
+                    result.Message = _configuration["EmployeeErrorMessage:getEmployeeErrorMessage"];
                     logger.LogError(result.Message, ex.ToString());
                 }
                 return result;
             }
-        }
-
+        } 
         //REMOVE
         public ServiceResult Remove(EmployeeDtoRemove DtoRemove)
         {
             ServiceResult result = new ServiceResult(); 
             try
             {
-                Employee employee = new Employee()
+                ServiceResult validation = _employeevalidations.DtoRemoveValidations(DtoRemove);
+                if (!validation.Success)
                 {
-                    EmpID = DtoRemove.EmpID,
-                    FirstName = DtoRemove.FirstName,
-                    LastName = DtoRemove.LastName,
-                    Minit = DtoRemove.Minit,
-                    Joblvl = DtoRemove.Joblvl,
-                    HireDate = DtoRemove.HireDate,
-                    CreationDate = DtoRemove.ChangeDate,
-                    IDCreationUser = DtoRemove.ChangeUser,
-                    Deleted = DtoRemove.Deleted
-                };
-
-                _employeeRepository.Remove(employee);
+                    return validation;
+                }
+                else
+                {
+                    Employee employee = new Employee()
+                    {
+                        EmpID = DtoRemove.EmpID,
+                        DeletedDate = DtoRemove.ChangeDate,
+                        DeletedUser = DtoRemove.ChangeUser,
+                        Deleted = DtoRemove.Deleted
+                    };
+                    _employeeRepository.Remove(employee);
+                    result.Message = _configuration["EmployeeSuccessMessages:removeEmployeeSuccessMessage"];
+                }
+            }
+            catch (EmployeeServiceExeption exx)
+            {
+                result.Success = false;
+                result.Message = exx.Message;
+                logger.LogError(result.Message, exx.Message);
             }
             catch (Exception ex)
             {
                 result.Success = false;
-                result.Message = $"An error has ocurred while removing the employee";
+                result.Message = _configuration["EmployeeErrorMessage:removedEmployeeErrorMessage"];
                 logger.LogError(result.Message, ex.ToString());
             }
             return result;
-        }
+        } 
 
         //SAVE
         public ServiceResult Save(EmployeeDtoAdd dtoadd)
@@ -123,28 +141,43 @@ namespace Publicaciones.Application.Service
             ServiceResult result = new ServiceResult();
             try
             {
-                Employee employee = new Employee()
+                ServiceResult validation = _employeevalidations.DtoAddValidations(dtoadd);
+                if (!validation.Success)
                 {
-                    CreationDate = dtoadd.ChangeDate,
-                    IDCreationUser = dtoadd.ChangeUser,
-                    FirstName = dtoadd.FirstName,
-                    LastName = dtoadd.LastName,
-                    HireDate = dtoadd.HireDate,
-                    Joblvl = dtoadd.Joblvl,
-                    Minit = dtoadd.Minit,
-                    JobID = dtoadd.JobID,
-                    PubID = dtoadd.PubID
-                };
-                _employeeRepository.Save(employee);
+                    return validation;
+                }
+                else
+                {
+                    Employee employee = new Employee()
+                    {
+                        CreationDate = dtoadd.ChangeDate,
+                        IDCreationUser = dtoadd.ChangeUser,
+                        FirstName = dtoadd.FirstName,
+                        LastName = dtoadd.LastName,
+                        HireDate = dtoadd.HireDate,
+                        Joblvl = dtoadd.Joblvl,
+                        Minit = dtoadd.Minit,
+                        JobID = dtoadd.JobID,
+                        PubID = dtoadd.PubID
+                    };
+                    _employeeRepository.Save(employee);
+                    result.Message = _configuration["EmployeeSuccessMessages:addEmployeeSuccessMessage"];
+                }
+            }
+            catch (EmployeeServiceExeption exx)
+            {
+                result.Success = false;
+                result.Message = exx.Message;
+                logger.LogError(result.Message, exx.Message);
             }
             catch (Exception ex)
             {
                 result.Success = false;
-                result.Message = $"The following error ocurred while saving the employee";
+                result.Message = _configuration["EmployeeErrorMessage:addEmployeeErrorMessage"];
                 logger.LogError(result.Message, ex.ToString());
             }
             return result;
-        }
+        } 
 
         //UPDATE
         public ServiceResult Update(EmployeeDtoUpdate dtoupdate)
@@ -152,28 +185,43 @@ namespace Publicaciones.Application.Service
             ServiceResult result = new ServiceResult();
             try
             {
-                Employee employee = new Employee()
+                ServiceResult validation = _employeevalidations.DtoUpdateValidations(dtoupdate);
+                if (!validation.Success)
                 {
-                    EmpID = dtoupdate.EmpID,
-                    CreationDate = dtoupdate.ChangeDate,
-                    IDCreationUser = dtoupdate.ChangeUser,
-                    FirstName = dtoupdate.FirstName,
-                    LastName = dtoupdate.LastName,
-                    HireDate = dtoupdate.HireDate,
-                    Joblvl = dtoupdate .Joblvl,
-                    Minit = dtoupdate.Minit,
-                    JobID = dtoupdate.JobID,
-                    PubID = dtoupdate.PubID
-                };
-                _employeeRepository.Update(employee);
+                    return validation;
+                }
+                else
+                {
+                    Employee employee = new Employee()
+                    {
+                        EmpID = dtoupdate.EmpID,
+                        CreationDate = dtoupdate.ChangeDate,
+                        IDCreationUser = dtoupdate.ChangeUser,
+                        FirstName = dtoupdate.FirstName,
+                        LastName = dtoupdate.LastName,
+                        HireDate = dtoupdate.HireDate,
+                        Joblvl = dtoupdate.Joblvl,
+                        Minit = dtoupdate.Minit,
+                        JobID = dtoupdate.JobID,
+                        PubID = dtoupdate.PubID
+                    };
+                    _employeeRepository.Update(employee);
+                    result.Message = _configuration["EmployeeSuccessMessages:updateEmployeeSuccessMessage"];
+                }
+            }
+            catch (EmployeeServiceExeption exx)
+            {
+                result.Success = false;
+                result.Message = exx.Message;
+                logger.LogError(result.Message, exx.Message);
             }
             catch (Exception ex)
             {
                 result.Success = false;
-                result.Message = $"An error has occured while updating the employee";
+                result.Message = _configuration["EmployeeErrorMessage:updateEmployeeErrorMessage"];
                 logger.LogError(result.Message, ex.ToString());
             }
             return result;
-        }
+        } 
     }
 }
